// example/src/theme/types.ts
export interface AppTheme {
  light: {
    primary: string;
    background: string;
    card: string;
    foreground: string;
    border: string;
    muted: string;
  };
  dark: {
    primary: string;
    background: string;
    card: string;
    foreground: string;
    border: string;
    muted: string;
  };
}